how to make new componenet files and export them
