--- src/App.js
@@ -1,26 +1,0 @@
-function App() {
-  return (
-    <div>
-      <Navbar />
-      <Footer />
-    </div>
-  );
-}
-
-function Navbar() {
-  return (
-    <div>
-      <h2>This is Navbar</h2>
-    </div>
-  );
-}
-
-function Footer() {
-  return (
-    <div>
-      <h3>This is Footer</h3>
-    </div>
-  );
-}
-
-export default App;
