+++ src/App.jsx
@@ -1,0 +1,13 @@
+import Navbar from "./components/navbar/Navbar";
+import Footer from "./components/footer/Footer";
+
+function App() {
+  return (
+    <div>
+      <Navbar />
+      <Footer />
+    </div>
+  );
+}
+
+export default App;
