+++ src/components/footer/Footer.jsx
@@ -1,0 +1,9 @@
+function Footer() {
+  return (
+    <div>
+      <h3>This is Footer</h3>
+    </div>
+  );
+}
+
+export default Footer;
