+++ src/components/navbar/Navbar.jsx
@@ -1,0 +1,9 @@
+function Navbar() {
+  return (
+    <div>
+      <h2>This is Navbar</h2>
+    </div>
+  );
+}
+
+export default Navbar;
